fix(agent): clean up host prompt lifecycle lint
